requirements.txt
# Core dependencies
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.4.2
torch>=2.3.0
imbalanced-learn>=0.12.0

# Data processing
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
pydantic>=2.7.0
pydantic-settings>=2.2.1

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# UI
streamlit>=1.36.0
plotly>=5.22.0
altair>=5.2.0

# Monitoring
prometheus-client>=0.20.0
mlflow>=2.13.1

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.1
joblib>=1.3.2
loguru>=0.7.2
tqdm>=4.66.4